tests: Move contact form tests into test_support_views.
The /request-demo/ and /support/ forms are implemented in
corporate/views/support.py, send email, and never touch Stripe or
billing; test_stripe is not where anyone would look for their tests.

Move them into test_support_views, based on where the views are
defined.  Without StripeTestCase's setUp, hamlet is an ordinary member
rather than a realm owner, so the support request email now reports
that role.
